Module winapi::shared

Headers shared between user mode and kernel mode

Modules

basetsd

Type definitions for the basic sized types.

cfg

common Configuration Manager definitions for both user mode and kernel mode code

devpropdef

Defines property types and keys for the Plug and Play Device Property API

guiddef

GUID definition

ktmtypes

Common types for KTM exposed at both the Nt- and Win32-layer

minwindef

Basic Windows Type Definitions for minwin partition

ntdef

Type definitions for the basic types.

ntstatus

Constant definitions for the NTSTATUS values.

rpcndr

windef

Basic Windows Type Definitions

winerror

error code definitions for the Win32 API functions

wtypesbase