Crate winapi

Types and constants for WinAPI bindings.

Reexports

pub use audioclient::*;

pub use basetsd::*;

pub use commctrl::*;

pub use d3d9::*;

pub use d3d9caps::*;

pub use d3d9types::*;

pub use dwmapi::*;

pub use fileapi::*;

pub use libloaderapi::*;

pub use minwinbase::*;

pub use minwindef::*;

pub use mmdeviceapi::*;

pub use mmreg::*;

pub use mmsystem::*;

pub use objidl::*;

pub use objidlbase::*;

pub use processthreadsapi::*;

pub use propsys::*;

pub use shobjidl::*;

pub use shtypes::*;

pub use synchapi::*;

pub use unknwnbase::*;

pub use vadefs::*;

pub use winbase::*;

pub use wincon::*;

pub use wincred::*;

pub use wincrypt::*;

pub use windowsx::*;

pub use winerror::*;

pub use wingdi::*;

pub use winioctl::*;

pub use winnetwk::*;

pub use winnls::*;

pub use winnt::*;

pub use winsvc::*;

pub use winuser::*;

pub use wtypesbase::*;

Modules

audioclient
this ALWAYS GENERATED file contains the definitions for the interfaces

basetsd
Type definitions for the basic sized types.

commctrl

d3d9
Direct3D include file

d3d9caps
Direct3D capabilities include file

d3d9types
Direct3D capabilities include file

dwmapi
Procedure declarations, constant definitions, and macros for the NLS component.

fileapi
ApiSet Contract for api-ms-win-core-file-l1

libloaderapi
ApiSet Contract for api-ms-win-core-libraryloader-l1

minwinbase
This module defines the 32-Bit Windows Base APIs

minwindef
Basic Windows Type Definitions for minwin partition

mmdeviceapi
this ALWAYS GENERATED file contains the definitions for the interfaces

mmreg

mmsystem

objidl
this ALWAYS GENERATED file contains the definitions for the interfaces

objidlbase
this ALWAYS GENERATED file contains the definitions for the interfaces

processthreadsapi

propsys

shobjidl
this ALWAYS GENERATED file contains the definitions for the interfaces

shtypes
this ALWAYS GENERATED file contains the definitions for the interfaces

synchapi
ApiSet Contract for api-ms-win-core-synch-l1

unknwnbase
this ALWAYS GENERATED file contains the definitions for the interfaces

vadefs
Definitions of macro helpers used by . This is the topmost header in the CRT header lattice, and is always the first CRT header to be included, explicitly or implicitly. Therefore, this header also has several definitions that are used throughout the CRT.

winbase
This module defines the 32-Bit Windows Base APIs

wincon
This module contains the public data structures, data types, and procedures exported by the NT console subsystem.

wincred
Authentication API Prototypes and Definitions

wincrypt
Cryptographic API Prototypes and Definitions

windowsx
Macro APIs, window message crackers, and control APIs

winerror
error code definitions for the Win32 API functions

wingdi
GDI procedure declarations, constant definitions and macros

winioctl
This module defines the 32-Bit Windows Device I/O control codes.

winnetwk
Standard WINNET Header File for WIN32

winnls
Procedure declarations, constant definitions, and macros for the NLS component.

winnt
This module defines the 32-Bit Windows types and constants that are defined by NT, but exposed through the Win32 API.

winsvc
Header file for the Service Control Manager

winuser
USER procedure declarations, constant definitions and macros

wtypesbase

Structs

CO_MTA_USAGE_COOKIE__

DEVMODEA

DEVMODEW

DISPLAY_DEVICEA

DISPLAY_DEVICEW

FLOAT128

GROUP_AFFINITY

GUID

HACCEL__

HBITMAP__

HBRUSH__

HCOLORSPACE__

HDC__

HDESK__

HENHMETAFILE__

HFONT__

HGLRC__

HHOOK__

HICON__

HMENU__

HMONITOR__

HPALETTE__

HPEN__

HUMPD__

HWINEVENTHOOK__

HWND__

IAccessibilityDockingService

IAccessibilityDockingServiceCallback

IAccessibleObject

IActionProgress

IActionProgressDialog

IAppVisibility

IAppVisibilityEvents

IApplicationActivationManager

IApplicationAssociationRegistration

IApplicationAssociationRegistrationUI

IApplicationDesignModeSettings

IApplicationDesignModeSettings2

IApplicationDestinations

IApplicationDocumentLists

IAssocHandler

IAssocHandlerInvoker

IAttachmentExecute

IAutoCompleteDropDown

IBandHost

IBandSite

IBrowserFrameOptions

ICDBurn

ICDBurnExt

ICategorizer

ICategoryProvider

IColumnManager

ICommDlgBrowser

ICommDlgBrowser2

ICommDlgBrowser3

IComputerInfoChangeNotify

IContactManagerInterop

IContextMenu

IContextMenu2

IContextMenu3

IContextMenuCB

IContextMenuSite

IControlMarkup

ICurrentItem

ICustomDestinationList

IDataObjectProvider

IDataTransferManagerInterop

IDefaultExtractIconInit

IDefaultFolderMenuInitialize

IDelegateFolder

IDelegateItem

IDeskBand

IDeskBand2

IDeskBandInfo

IDeskBar

IDesktopGadget

IDesktopWallpaper

IDestinationStreamFactory

IDisplayItem

IDockingWindow

IDragSourceHelper

IDragSourceHelper2

IDropTargetHelper

IDynamicHWHandler

IEnumAssocHandlers

IEnumExplorerCommand

IEnumExtraSearch

IEnumFullIDList

IEnumIDList

IEnumObjects

IEnumReadyCallback

IEnumResources

IEnumShellItems

IEnumerableView

IExecuteCommand

IExecuteCommandApplicationHostEnvironment

IExecuteCommandHost

IExplorerBrowser

IExplorerBrowserEvents

IExplorerCommand

IExplorerCommandProvider

IExplorerCommandState

IExplorerPaneVisibility

IExtractImage

IExtractImage2

IFileIsInUse

IFileOperation

IFileSyncMergeHandler

IFileSystemBindData

IFileSystemBindData2

IFolderBandPriv

IFolderFilter

IFolderFilterSite

IFolderView

IFolderView2

IFolderViewHost

IFolderViewOptions

IFolderViewSettings

IFrameworkInputPane

IFrameworkInputPaneHandler

IHWEventHandler

IHWEventHandler2

IHandlerActivationHost

IHandlerInfo

IHomeGroup

IIOCancelInformation

IIdentityName

IImageRecompress

IInitializeCommand

IInitializeNetworkFolder

IInitializeWithBindCtx

IInitializeWithItem

IInitializeWithPropertyStore

IInitializeWithWindow

IInputObject

IInputObject2

IInputObjectSite

IInsertItem

IItemNameLimits

IKnownFolder

IKnownFolderManager

ILaunchSourceAppUserModelId

ILaunchSourceViewSizePreference

ILaunchTargetMonitor

ILaunchTargetViewSizePreference

IMarkupCallback

IMenuBand

IMenuPopup

INameSpaceTreeAccessible

INameSpaceTreeControl

INameSpaceTreeControl2

INameSpaceTreeControlCustomDraw

INameSpaceTreeControlDropHandler

INameSpaceTreeControlEvents

INameSpaceTreeControlFolderCapabilities

INamespaceWalk

INamespaceWalkCB

INamespaceWalkCB2

INewMenuClient

INewWindowManager

IObjectProvider

IObjectWithAppUserModelID

IObjectWithBackReferences

IObjectWithCancelEvent

IObjectWithFolderEnumMode

IObjectWithProgID

IObjectWithSelection

IOpenControlPanel

IOpenSearchSource

IOperationsProgressDialog

IPackageDebugSettings

IPackageExecutionStateChangeNotification

IParentAndItem

IParseAndCreateItem

IPersistFolder

IPersistFolder2

IPersistFolder3

IPersistIDList

IPreviewHandler

IPreviewHandlerFrame

IPreviewHandlerVisuals

IPreviewItem

IPreviousVersionsInfo

IProfferService

IPropertyUI

IPublishingWizard

IQueryCancelAutoPlay

IQueryCodePage

IQueryContinue

IRegTreeItem

IRelatedItem

IRemoteComputer

IResolveShellLink

IResultsFolder

IRunnableTask

ISearchBoxInfo

ISearchFolderItemFactory

ISharingConfigurationManager

IShellBrowser

IShellExtInit

IShellFolder

IShellFolder2

IShellIcon

IShellItem2

IShellItemImageFactory

IShellItemResources

IShellLibrary

IShellLinkA

IShellLinkDataList

IShellLinkW

IShellMenu

IShellMenuCallback

IShellPropSheetExt

IShellRunDll

IShellTaskScheduler

IShellView

IShellView2

IShellView3

IStartMenuPinnedList

IStreamAsync

IStreamUnbufferedInfo

ISuspensionDependencyManager

ITaskbarList

ITaskbarList2

ITaskbarList3

ITaskbarList4

IThumbnailHandlerFactory

ITransferAdviseSink

ITransferDestination

ITransferMediumItem

ITransferSource

ITrayDeskBand

IUpdateIDList

IUseToBrowseItem

IUserAccountChangeCallback

IUserNotification

IUserNotification2

IUserNotificationCallback

IViewStateIdentityItem

IVisualProperties

IWebWizardExtension

IWizardExtension

IWizardSite

LIST_ENTRY

LIST_ENTRY32

LIST_ENTRY64

LUID

LUID_AND_ATTRIBUTES

M128A

MSG

OBJECTID

PAINTSTRUCT

PIXELFORMATDESCRIPTOR

POINT

POINTL

PROCESSOR_NUMBER

PROCESS_MEMORY_COUNTERS

PROCESS_MEMORY_COUNTERS_EX

PROPVARIANT

RECT

SCROLLBARINFO

SCROLLINFO

SINGLE_LIST_ENTRY

SIZE

ServerInformation

TOKEN_PRIVILEGES

VALENTA

VALENTW

WAVEFORMATEX

WINDOWPLACEMENT

WNDCLASSEXW

WNDCLASSW

XSAVE_FORMAT

_CONTEXT

_DISPATCHER_CONTEXT

_EXCEPTION_RECORD

Enums

AUDCLNT_SHAREMODE

COMPARTMENT_ID

EXCEPTION_DISPOSITION

KNOWN_FOLDER_FLAG

c_void
Type used to construct void pointers for use with C.

Constants

BN_CLICKED

BN_DBLCLK

BN_DISABLE

BN_DOUBLECLICKED

BN_HILITE

BN_KILLFOCUS

BN_PAINT

BN_PUSHED

BN_SETFOCUS

BN_UNHILITE

BN_UNPUSHED

BS_3STATE

BS_AUTO3STATE

BS_AUTOCHECKBOX

BS_AUTORADIOBUTTON

BS_BITMAP

BS_BOTTOM

BS_CENTER

BS_CHECKBOX

BS_DEFPUSHBUTTON

BS_FLAT

BS_GROUPBOX

BS_ICON

BS_LEFT

BS_LEFTTEXT

BS_MULTILINE

BS_NOTIFY

BS_OWNERDRAW

BS_PUSHBOX

BS_PUSHBUTTON

BS_PUSHLIKE

BS_RADIOBUTTON

BS_RIGHT

BS_RIGHTBUTTON

BS_TEXT

BS_TOP

BS_TYPEMASK

BS_USERBUTTON

BS_VCENTER

CCHFORMNAME

CCHILDREN_SCROLLBAR

CDS_DISABLE_UNSAFE_MODES

CDS_ENABLE_UNSAFE_MODES

CDS_FULLSCREEN

CDS_GLOBAL

CDS_NORESET

CDS_RESET

CDS_RESET_EX

CDS_SET_PRIMARY

CDS_TEST

CDS_UPDATEREGISTRY

CDS_VIDEOPARAMETERS

CLSCTX_ALL

CLSCTX_INPROC_HANDLER

CLSCTX_INPROC_SERVER

CLSCTX_LOCAL_SERVER

CLSCTX_REMOTE_SERVER

CLSCTX_SERVER

CS_BYTEALIGNCLIENT

CS_BYTEALIGNWINDOW

CS_CLASSDC

CS_DBLCLKS

CS_DROPSHADOW

CS_GLOBALCLASS

CS_HREDRAW

CS_IME

CS_NOCLOSE

CS_OWNDC

CS_PARENTDC

CS_SAVEBITS

CS_VREDRAW

CW_USEDEFAULT

DISPLAY_DEVICE_ACC_DRIVER

DISPLAY_DEVICE_ACTIVE

DISPLAY_DEVICE_ATTACHED

DISPLAY_DEVICE_ATTACHED_TO_DESKTOP

DISPLAY_DEVICE_DISCONNECT

DISPLAY_DEVICE_MIRRORING_DRIVER

DISPLAY_DEVICE_MODESPRUNED

DISPLAY_DEVICE_MULTI_DRIVER

DISPLAY_DEVICE_PRIMARY_DEVICE

DISPLAY_DEVICE_REMOTE

DISPLAY_DEVICE_REMOVABLE

DISPLAY_DEVICE_TS_COMPATIBLE

DISPLAY_DEVICE_UNSAFE_MODES_ON

DISPLAY_DEVICE_VGA_COMPATIBLE

DISP_CHANGE_BADDUALVIEW

DISP_CHANGE_BADFLAGS

DISP_CHANGE_BADMODE

DISP_CHANGE_BADPARAM

DISP_CHANGE_FAILED

DISP_CHANGE_NOTUPDATED

DISP_CHANGE_RESTART

DISP_CHANGE_SUCCESSFUL

DM_BITSPERPEL

DM_COLLATE

DM_COLOR

DM_COPIES

DM_DEFAULTSOURCE

DM_DISPLAYFIXEDOUTPUT

DM_DISPLAYFLAGS

DM_DISPLAYFREQUENCY

DM_DISPLAYORIENTATION

DM_DITHERTYPE

DM_DUPLEX

DM_FORMNAME

DM_ICMINTENT

DM_ICMMETHOD

DM_LOGPIXELS

DM_MEDIATYPE

DM_NUP

DM_ORIENTATION

DM_PANNINGHEIGHT

DM_PANNINGWIDTH

DM_PAPERLENGTH

DM_PAPERSIZE

DM_PAPERWIDTH

DM_PELSHEIGHT

DM_PELSWIDTH

DM_POSITION

DM_PRINTQUALITY

DM_SCALE

DM_TTOPTION

DM_YRESOLUTION

EDD_GET_DEVICE_INTERFACE_NAME

ENUM_CURRENT_SETTINGS

ENUM_REGISTRY_SETTINGS

GPFIDL_ALTNAME

GPFIDL_DEFAULT

GPFIDL_UNCPRINTER

GW_CHILD

GW_ENABLEDPOPUP

GW_HWNDFIRST

GW_HWNDLAST

GW_HWNDNEXT

GW_HWNDPREV

GW_MAX

GW_OWNER

HKEY_CLASSES_ROOT

HKEY_CURRENT_CONFIG

HKEY_CURRENT_USER

HKEY_CURRENT_USER_LOCAL_SETTINGS

HKEY_DYN_DATA

HKEY_LOCAL_MACHINE

HKEY_PERFORMANCE_DATA

HKEY_PERFORMANCE_NLSTEXT

HKEY_PERFORMANCE_TEXT

HKEY_USERS

IDO_SHGIOI_DEFAULT

IDO_SHGIOI_LINK

IDO_SHGIOI_SHARE

IDO_SHGIOI_SLOWFILE

INVALID_HANDLE_VALUE

KEY_ALL_ACCESS

KEY_CREATE_LINK

KEY_CREATE_SUB_KEY

KEY_ENUMERATE_SUB_KEYS

KEY_EXECUTE

KEY_NOTIFY

KEY_QUERY_VALUE

KEY_READ

KEY_SET_VALUE

KEY_WOW64_32KEY

KEY_WOW64_64KEY

KEY_WOW64_RES

KEY_WRITE

LSFW_LOCK

LSFW_UNLOCK

MB_ABORTRETRYIGNORE

MB_APPLMODAL

MB_CANCELTRYCONTINUE

MB_DEFAULT_DESKTOP_ONLY

MB_DEFBUTTON1

MB_DEFBUTTON2

MB_DEFBUTTON3

MB_DEFBUTTON4

MB_DEFMASK

MB_HELP

MB_ICONASTERISK

MB_ICONERROR

MB_ICONEXCLAMATION

MB_ICONHAND

MB_ICONINFORMATION

MB_ICONMASK

MB_ICONQUESTION

MB_ICONSTOP

MB_ICONWARNING

MB_MISCMASK

MB_MODEMASK

MB_NOFOCUS

MB_OK

MB_OKCANCEL

MB_RETRYCANCEL

MB_RIGHT

MB_RTLREADING

MB_SERVICE_NOTIFICATION

MB_SERVICE_NOTIFICATION_NT3X

MB_SETFOREGROUND

MB_SYSTEMMODAL

MB_TASKMODAL

MB_TOPMOST

MB_TYPEMASK

MB_USERICON

MB_YESNO

MB_YESNOCANCEL

MDITILE_HORIZONTAL

MDITILE_SKIPDISABLED

MDITILE_VERTICAL

MDITILE_ZORDER

OFASI_EDIT

OFASI_OPENDESKTOP

PFD_DEPTH_DONTCARE

PFD_DIRECT3D_ACCELERATED

PFD_DOUBLEBUFFER

PFD_DOUBLEBUFFER_DONTCARE

PFD_DRAW_TO_BITMAP

PFD_DRAW_TO_WINDOW

PFD_GENERIC_ACCELERATED

PFD_GENERIC_FORMAT

PFD_MAIN_PLANE

PFD_NEED_PALETTE

PFD_NEED_SYSTEM_PALETTE

PFD_OVERLAY_PLANE

PFD_STEREO

PFD_STEREO_DONTCARE

PFD_SUPPORT_COMPOSITION

PFD_SUPPORT_DIRECTDRAW

PFD_SUPPORT_GDI

PFD_SUPPORT_OPENGL

PFD_SWAP_COPY

PFD_SWAP_EXCHANGE

PFD_SWAP_LAYER_BUFFERS

PFD_TYPE_COLORINDEX

PFD_TYPE_RGBA

PFD_UNDERLAY_PLANE

PROCESS_ALL_ACCESS

PROCESS_CREATE_PROCESS

PROCESS_CREATE_THREAD

PROCESS_DUP_HANDLE

PROCESS_QUERY_INFORMATION

PROCESS_QUERY_LIMITED_INFORMATION

PROCESS_SET_INFORMATION

PROCESS_SET_LIMITED_INFORMATION

PROCESS_SET_QUOTA

PROCESS_SET_SESSIONID

PROCESS_SUSPEND_RESUME

PROCESS_TERMINATE

PROCESS_VM_OPERATION

PROCESS_VM_READ

PROCESS_VM_WRITE

REG_BINARY

REG_CREATED_NEW_KEY

REG_DWORD

REG_DWORD_BIG_ENDIAN

REG_DWORD_LITTLE_ENDIAN

REG_EXPAND_SZ

REG_FULL_RESOURCE_DESCRIPTOR

REG_LEGAL_CHANGE_FILTER

REG_LINK

REG_MUI_STRING_TRUNCATE

REG_MULTI_SZ

REG_NONE

REG_NOTIFY_CHANGE_ATTRIBUTES

REG_NOTIFY_CHANGE_LAST_SET

REG_NOTIFY_CHANGE_NAME

REG_NOTIFY_CHANGE_SECURITY

REG_NOTIFY_THREAD_AGNOSTIC

REG_OPENED_EXISTING_KEY

REG_OPTION_BACKUP_RESTORE

REG_OPTION_CREATE_LINK

REG_OPTION_NON_VOLATILE

REG_OPTION_OPEN_LINK

REG_OPTION_RESERVED

REG_OPTION_VOLATILE

REG_QWORD

REG_QWORD_LITTLE_ENDIAN

REG_RESOURCE_LIST

REG_RESOURCE_REQUIREMENTS_LIST

REG_SZ

RRF_NOEXPAND

RRF_RT_ANY

RRF_RT_DWORD

RRF_RT_QWORD

RRF_RT_REG_BINARY

RRF_RT_REG_DWORD

RRF_RT_REG_EXPAND_SZ

RRF_RT_REG_MULTI_SZ

RRF_RT_REG_NONE

RRF_RT_REG_QWORD

RRF_RT_REG_SZ

RRF_ZEROONFAILURE

SB_BOTH

SB_CTL

SB_HORZ

SB_VERT

SND_ALIAS

SND_ALIAS_ID

SND_APPLICATION

SND_ASYNC

SND_FILENAME

SND_LOOP

SND_MEMORY

SND_NODEFAULT

SND_NOSTOP

SND_NOWAIT

SND_PURGE

SND_RESOURCE

SND_RING

SND_SENTRY

SND_SYNC

SND_SYSTEM

SWP_ASYNCWINDOWPOS

SWP_DEFERERASE

SWP_DRAWFRAME

SWP_FRAMECHANGED

SWP_HIDEWINDOW

SWP_NOACTIVATE

SWP_NOCOPYBITS

SWP_NOMOVE

SWP_NOOWNERZORDER

SWP_NOREDRAW

SWP_NOREPOSITION

SWP_NOSENDCHANGING

SWP_NOSIZE

SWP_NOZORDER

SWP_SHOWWINDOW

SW_FORCEMINIMIZE

SW_HIDE

SW_MAX

SW_MAXIMIZE

SW_MINIMIZE

SW_NORMAL

SW_RESTORE

SW_SHOW

SW_SHOWDEFAULT

SW_SHOWMAXIMIZED

SW_SHOWMINIMIZED

SW_SHOWMINNOACTIVE

SW_SHOWNA

SW_SHOWNOACTIVATE

SW_SHOWNORMAL

TOKEN_ADJUST_DEFAULT

TOKEN_ADJUST_GROUPS

TOKEN_ADJUST_PRIVILEGES

TOKEN_ADJUST_SESSIONID

TOKEN_ALL_ACCESS

TOKEN_ALL_ACCESS_P

TOKEN_ASSIGN_PRIMARY

TOKEN_DUPLICATE

TOKEN_EXECUTE

TOKEN_IMPERSONATE

TOKEN_QUERY

TOKEN_QUERY_SOURCE

TOKEN_READ

TOKEN_WRITE

VK_ACCEPT

VK_ADD

VK_APPS

VK_ATTN

VK_BACK

VK_BROWSER_BACK

VK_BROWSER_FAVORITES

VK_BROWSER_FORWARD

VK_BROWSER_HOME

VK_BROWSER_REFRESH

VK_BROWSER_SEARCH

VK_BROWSER_STOP

VK_CANCEL

VK_CAPITAL

VK_CLEAR

VK_CONTROL

VK_CONVERT

VK_CRSEL

VK_DECIMAL

VK_DELETE

VK_DIVIDE

VK_DOWN

VK_END

VK_EREOF

VK_ESCAPE

VK_EXECUTE

VK_EXSEL

VK_F1

VK_F10

VK_F11

VK_F12

VK_F13

VK_F14

VK_F15

VK_F16

VK_F17

VK_F18

VK_F19

VK_F2

VK_F20

VK_F21

VK_F22

VK_F23

VK_F24

VK_F3

VK_F4

VK_F5

VK_F6

VK_F7

VK_F8

VK_F9

VK_FINAL

VK_HANGUEL

VK_HANGUL

VK_HANJA

VK_HELP

VK_HOME

VK_ICO_00

VK_ICO_CLEAR

VK_ICO_HELP

VK_INSERT

VK_JUNJA

VK_KANA

VK_KANJI

VK_LAUNCH_APP1

VK_LAUNCH_APP2

VK_LAUNCH_MAIL

VK_LAUNCH_MEDIA_SELECT

VK_LBUTTON

VK_LCONTROL

VK_LEFT

VK_LMENU

VK_LSHIFT

VK_LWIN

VK_MBUTTON

VK_MEDIA_NEXT_TRACK

VK_MEDIA_PLAY_PAUSE

VK_MEDIA_PREV_TRACK

VK_MEDIA_STOP

VK_MENU

VK_MODECHANGE

VK_MULTIPLY

VK_NEXT

VK_NONAME

VK_NONCONVERT

VK_NUMLOCK

VK_NUMPAD0

VK_NUMPAD1

VK_NUMPAD2

VK_NUMPAD3

VK_NUMPAD4

VK_NUMPAD5

VK_NUMPAD6

VK_NUMPAD7

VK_NUMPAD8

VK_NUMPAD9

VK_OEM_1

VK_OEM_102

VK_OEM_2

VK_OEM_3

VK_OEM_4

VK_OEM_5

VK_OEM_6

VK_OEM_7

VK_OEM_8

VK_OEM_ATTN

VK_OEM_AUTO

VK_OEM_AX

VK_OEM_BACKTAB

VK_OEM_CLEAR

VK_OEM_COMMA

VK_OEM_COPY

VK_OEM_CUSEL

VK_OEM_ENLW

VK_OEM_FINISH

VK_OEM_FJ_JISHO

VK_OEM_FJ_LOYA

VK_OEM_FJ_MASSHOU

VK_OEM_FJ_ROYA

VK_OEM_FJ_TOUROKU

VK_OEM_JUMP

VK_OEM_MINUS

VK_OEM_NEC_EQUAL

VK_OEM_PA1

VK_OEM_PA2

VK_OEM_PA3

VK_OEM_PERIOD

VK_OEM_PLUS

VK_OEM_RESET

VK_OEM_WSCTRL

VK_PA1

VK_PACKET

VK_PAUSE

VK_PLAY

VK_PRINT

VK_PRIOR

VK_PROCESSKEY

VK_RBUTTON

VK_RCONTROL

VK_RETURN

VK_RIGHT

VK_RMENU

VK_RSHIFT

VK_RWIN

VK_SCROLL

VK_SELECT

VK_SEPARATOR

VK_SHIFT

VK_SLEEP

VK_SNAPSHOT

VK_SPACE

VK_SUBTRACT

VK_TAB

VK_UP

VK_VOLUME_DOWN

VK_VOLUME_MUTE

VK_VOLUME_UP

VK_XBUTTON1

VK_XBUTTON2

VK_ZOOM

WM_ACTIVATE

WM_ACTIVATEAPP

WM_AFXFIRST

WM_AFXLAST

WM_APP

WM_APPCOMMAND

WM_ASKCBFORMATNAME

WM_CANCELJOURNAL

WM_CANCELMODE

WM_CAPTURECHANGED

WM_CHANGECBCHAIN

WM_CHANGEUISTATE

WM_CHAR

WM_CHARTOITEM

WM_CHILDACTIVATE

WM_CLEAR

WM_CLIPBOARDUPDATE

WM_CLOSE

WM_COMMAND

WM_COMMNOTIFY

WM_COMPACTING

WM_COMPAREITEM

WM_CONTEXTMENU

WM_COPY

WM_COPYDATA

WM_CREATE

WM_CTLCOLORBTN

WM_CTLCOLORDLG

WM_CTLCOLOREDIT

WM_CTLCOLORLISTBOX

WM_CTLCOLORMSGBOX

WM_CTLCOLORSCROLLBAR

WM_CTLCOLORSTATIC

WM_CUT

WM_DEADCHAR

WM_DELETEITEM

WM_DESTROY

WM_DESTROYCLIPBOARD

WM_DEVICECHANGE

WM_DEVMODECHANGE

WM_DISPLAYCHANGE

WM_DPICHANGED

WM_DRAWCLIPBOARD

WM_DRAWITEM

WM_DROPFILES

WM_DWMCOLORIZATIONCOLORCHANGED

WM_DWMCOMPOSITIONCHANGED

WM_DWMNCRENDERINGCHANGED

WM_DWMSENDICONICLIVEPREVIEWBITMAP

WM_DWMSENDICONICTHUMBNAIL

WM_DWMWINDOWMAXIMIZEDCHANGE

WM_ENABLE

WM_ENDSESSION

WM_ENTERIDLE

WM_ENTERMENULOOP

WM_ENTERSIZEMOVE

WM_ERASEBKGND

WM_EXITMENULOOP

WM_EXITSIZEMOVE

WM_FONTCHANGE

WM_GESTURE

WM_GESTURENOTIFY

WM_GETDLGCODE

WM_GETFONT

WM_GETHOTKEY

WM_GETICON

WM_GETMINMAXINFO

WM_GETOBJECT

WM_GETTEXT

WM_GETTEXTLENGTH

WM_GETTITLEBARINFOEX

WM_HANDHELDFIRST

WM_HANDHELDLAST

WM_HELP

WM_HOTKEY

WM_HSCROLL

WM_HSCROLLCLIPBOARD

WM_ICONERASEBKGND

WM_IME_CHAR

WM_IME_COMPOSITION

WM_IME_COMPOSITIONFULL

WM_IME_CONTROL

WM_IME_ENDCOMPOSITION

WM_IME_KEYDOWN

WM_IME_KEYLAST

WM_IME_KEYUP

WM_IME_NOTIFY

WM_IME_REQUEST

WM_IME_SELECT

WM_IME_SETCONTEXT

WM_IME_STARTCOMPOSITION

WM_INITDIALOG

WM_INITMENU

WM_INITMENUPOPUP

WM_INPUT

WM_INPUTLANGCHANGE

WM_INPUTLANGCHANGEREQUEST

WM_INPUT_DEVICE_CHANGE

WM_KEYDOWN

WM_KEYFIRST

WM_KEYLAST

WM_KEYUP

WM_KILLFOCUS

WM_LBUTTONDBLCLK

WM_LBUTTONDOWN

WM_LBUTTONUP

WM_MBUTTONDBLCLK

WM_MBUTTONDOWN

WM_MBUTTONUP

WM_MDIACTIVATE

WM_MDICASCADE

WM_MDICREATE

WM_MDIDESTROY

WM_MDIGETACTIVE

WM_MDIICONARRANGE

WM_MDIMAXIMIZE

WM_MDINEXT

WM_MDIREFRESHMENU

WM_MDIRESTORE

WM_MDISETMENU

WM_MDITILE

WM_MEASUREITEM

WM_MENUCHAR

WM_MENUCOMMAND

WM_MENUDRAG

WM_MENUGETOBJECT

WM_MENURBUTTONUP

WM_MENUSELECT

WM_MOUSEACTIVATE

WM_MOUSEFIRST

WM_MOUSEHOVER

WM_MOUSEHWHEEL

WM_MOUSELAST

WM_MOUSELEAVE

WM_MOUSEMOVE

WM_MOUSEWHEEL

WM_MOVE

WM_MOVING

WM_NCACTIVATE

WM_NCCALCSIZE

WM_NCCREATE

WM_NCDESTROY

WM_NCHITTEST

WM_NCLBUTTONDBLCLK

WM_NCLBUTTONDOWN

WM_NCLBUTTONUP

WM_NCMBUTTONDBLCLK

WM_NCMBUTTONDOWN

WM_NCMBUTTONUP

WM_NCMOUSEHOVER

WM_NCMOUSELEAVE

WM_NCMOUSEMOVE

WM_NCPAINT

WM_NCPOINTERDOWN

WM_NCPOINTERUP

WM_NCPOINTERUPDATE

WM_NCRBUTTONDBLCLK

WM_NCRBUTTONDOWN

WM_NCRBUTTONUP

WM_NCXBUTTONDBLCLK

WM_NCXBUTTONDOWN

WM_NCXBUTTONUP

WM_NEXTDLGCTL

WM_NEXTMENU

WM_NOTIFY

WM_NOTIFYFORMAT

WM_NULL

WM_PAINT

WM_PAINTCLIPBOARD

WM_PAINTICON

WM_PALETTECHANGED

WM_PALETTEISCHANGING

WM_PARENTNOTIFY

WM_PASTE

WM_PENWINFIRST

WM_PENWINLAST

WM_POINTERACTIVATE

WM_POINTERCAPTURECHANGED

WM_POINTERDEVICECHANGE

WM_POINTERDEVICEINRANGE

WM_POINTERDEVICEOUTOFRANGE

WM_POINTERDOWN

WM_POINTERENTER

WM_POINTERHWHEEL

WM_POINTERLEAVE

WM_POINTERUP

WM_POINTERUPDATE

WM_POINTERWHEEL

WM_POWER

WM_POWERBROADCAST

WM_PRINT

WM_PRINTCLIENT

WM_QUERYDRAGICON

WM_QUERYENDSESSION

WM_QUERYNEWPALETTE

WM_QUERYOPEN

WM_QUERYUISTATE

WM_QUEUESYNC

WM_QUIT

WM_RBUTTONDBLCLK

WM_RBUTTONDOWN

WM_RBUTTONUP

WM_RENDERALLFORMATS

WM_RENDERFORMAT

WM_SETCURSOR

WM_SETFOCUS

WM_SETFONT

WM_SETHOTKEY

WM_SETICON

WM_SETREDRAW

WM_SETTEXT

WM_SETTINGCHANGE

WM_SHOWWINDOW

WM_SIZE

WM_SIZECLIPBOARD

WM_SIZING

WM_SPOOLERSTATUS

WM_STYLECHANGED

WM_STYLECHANGING

WM_SYNCPAINT

WM_SYSCHAR

WM_SYSCOLORCHANGE

WM_SYSCOMMAND

WM_SYSDEADCHAR

WM_SYSKEYDOWN

WM_SYSKEYUP

WM_TABLET_FIRST

WM_TABLET_LAST

WM_TCARD

WM_THEMECHANGED

WM_TIMECHANGE

WM_TIMER

WM_TOUCH

WM_TOUCHHITTESTING

WM_UNDO

WM_UNICHAR

WM_UNINITMENUPOPUP

WM_UPDATEUISTATE

WM_USER

WM_USERCHANGED

WM_VKEYTOITEM

WM_VSCROLL

WM_VSCROLLCLIPBOARD

WM_WINDOWPOSCHANGED

WM_WINDOWPOSCHANGING

WM_WININICHANGE

WM_WTSSESSION_CHANGE

WM_XBUTTONDBLCLK

WM_XBUTTONDOWN

WM_XBUTTONUP

WS_BORDER

WS_CAPTION

WS_CHILD

WS_CHILDWINDOW

WS_CLIPCHILDREN

WS_CLIPSIBLINGS

WS_DISABLED

WS_DLGFRAME

WS_EX_ACCEPTFILES

WS_EX_APPWINDOW

WS_EX_CLIENTEDGE

WS_EX_COMPOSITED

WS_EX_CONTEXTHELP

WS_EX_CONTROLPARENT

WS_EX_DLGMODALFRAME

WS_EX_LAYERED

WS_EX_LAYOUTRTL

WS_EX_LEFT

WS_EX_LEFTSCROLLBAR

WS_EX_LTRREADING

WS_EX_MDICHILD

WS_EX_NOACTIVATE

WS_EX_NOINHERITLAYOUT

WS_EX_NOPARENTNOTIFY

WS_EX_NOREDIRECTIONBITMAP

WS_EX_OVERLAPPEDWINDOW

WS_EX_PALETTEWINDOW

WS_EX_RIGHT

WS_EX_RIGHTSCROLLBAR

WS_EX_RTLREADING

WS_EX_STATICEDGE

WS_EX_TOOLWINDOW

WS_EX_TOPMOST

WS_EX_TRANSPARENT

WS_EX_WINDOWEDGE

WS_GROUP

WS_HSCROLL

WS_ICONIC

WS_MAXIMIZE

WS_MAXIMIZEBOX

WS_MINIMIZE

WS_MINIMIZEBOX

WS_OVERLAPPED

WS_OVERLAPPEDWINDOW

WS_POPUP

WS_POPUPWINDOW

WS_SIZEBOX

WS_SYSMENU

WS_TABSTOP

WS_THICKFRAME

WS_TILED

WS_TILEDWINDOW

WS_VISIBLE

WS_VSCROLL

Type Definitions

ACCESS_MASK

BOOLEAN

CCHAR

CLSID

COLORREF

CO_MTA_USAGE_COOKIE

DESKTOPENUMPROCA

DESKTOPENUMPROCW

DWORDLONG

FCHAR

FLONG

FMTID

FSHORT

GPFIDL_FLAGS

HACCEL

HANDLE

HBITMAP

HBRUSH

HCOLORSPACE

HCURSOR

HDC

HDESK

HDEVNOTIFY

HDWP

HENHMETAFILE

HFONT

HGDIOBJ

HGLRC

HHOOK

HICON

HMENU

HMONITOR

HOOKPROC

HPALETTE

HPEN

HRESULT

HUMPD

HWINEVENTHOOK

HWND

IID

KSPIN_LOCK

LANGID

LARGE_INTEGER

LCID

LONGLONG

LPCCH

LPCGUID

LPCH

LPCLSID

LPCOLORREF

LPCRECT

LPCSCROLLINFO

LPCSTR

LPCUWCHAR

LPCUWSTR

LPCWCH

LPCWCHAR

LPCWSTR

LPDEVMODEA

LPDEVMODEW

LPDISPLAY_DEVICEA

LPDISPLAY_DEVICEW

LPFMTID

LPGUID

LPIID

LPMSG

LPPAINTSTRUCT

LPPIXELFORMATDESCRIPTOR

LPPOINT

LPRECT

LPSCROLLBARINFO

LPSCROLLINFO

LPSTR

LPUWSTR

LPWCH

LPWINDOWPLACEMENT

LPWNDCLASSEXW

LPWNDCLASSW

LPWSTR

NAMEENUMPROCA

NAMEENUMPROCW

NPDEVMODEA

NPDEVMODEW

NPMSG

NPPAINTSTRUCT

NPPOINT

NPRECT

NPSTR

NPWNDCLASSEXW

NPWNDCLASSW

NWPSTR

PACCESS_MASK

PBOOLEAN

PCCH

PCH

PCHAR

PCNZCH

PCNZWCH

PCOMPARTMENT_ID

PCSTR

PCUCSCHAR

PCUCSSTR

PCUNZWCH

PCUUCSCHAR

PCUUCSSTR

PCUWCHAR

PCUWSTR

PCUZZWSTR

PCWCH

PCWCHAR

PCWSTR

PCZPCSTR

PCZPCWSTR

PCZPSTR

PCZPWSTR

PCZZSTR

PCZZWSTR

PDEVMODEA

PDEVMODEW

PDISPLAY_DEVICEA

PDISPLAY_DEVICEW

PDWORDLONG

PEXCEPTION_ROUTINE

PFLOAT128

PGROUP_AFFINITY

PHANDLE

PKSPIN_LOCK

PLARGE_INTEGER

PLCID

PLIST_ENTRY

PLIST_ENTRY32

PLIST_ENTRY64

PLONG

PLONGLONG

PLUID

PLUID_AND_ATTRIBUTES

PM128A

PMSG

PNZCH

PNZWCH

PPAINTSTRUCT

PPIXELFORMATDESCRIPTOR

PPOINT

PPOINTL

PPROCESSOR_NUMBER

PPROCESS_MEMORY_COUNTERS

PRECT

PRTL_REFERENCE_COUNT

PSCROLLBARINFO

PSHORT

PSINGLE_LIST_ENTRY

PSIZE

PSTR

PServerInformation

PTOKEN_PRIVILEGES

PUCSCHAR

PUCSSTR

PULARGE_INTEGER

PULONGLONG

PUNZWCH

PUUCSCHAR

PUUCSSTR

PUWSTR

PUZZWSTR

PVALENTA

PVALENTW

PVOID

PVOID64

PWCH

PWCHAR

PWINDOWPLACEMENT

PWNDCLASSEXW

PWNDCLASSW

PWSTR

PZPCSTR

PZPCWSTR

PZPSTR

PZPWSTR

PZZSTR

PZZWSTR

REFCLSID

REFERENCE_TIME

REFFMTID

REFGUID

REFIID

REGSAM

RTL_REFERENCE_COUNT

TimerProc

UCSCHAR

ULARGE_INTEGER

ULONGLONG

USN

VARTYPE

WCHAR

WNDENUMPROC

WNDPROC

__int16

__int32

__int64

__int8

__uint16

__uint32

__uint64

__uint8

c_char

c_double

c_float

c_int

c_long

c_longlong

c_schar

c_short

c_uchar

c_uint

c_ulong

c_ulonglong

c_ushort

wchar_t