Crate winapi

Modules

ctypes
Built in primitive types provided by the C language

shared
Headers shared between user mode and kernel mode

um
Headers for user mode only

vc
Headers that come with VC. Notably, these are not part of the Windows SDK.

winrt

Macros

DEFINE_DEVPROPKEY

DEFINE_GUID

DEFINE_PROPERTYKEY

ENUM

RIDL

STRUCT

Traits

Class

Interface