Crate winapi

Types and constants for WinAPI bindings.

Reexports

pub use audioclient::*;

pub use audiosessiontypes::*;

pub use basetsd::*;

pub use bcrypt::*;

pub use cfg::*;

pub use cfgmgr32::*;

pub use commctrl::*;

pub use commdlg::*;

pub use corsym::*;

pub use d2d1::*;

pub use d2dbasetypes::*;

pub use d3d9::*;

pub use d3d9caps::*;

pub use d3d9types::*;

pub use d3d11::*;

pub use d3d12::*;

pub use d3d12sdklayers::*;

pub use d3dcommon::*;

pub use dbghelp::*;

pub use dcommon::*;

pub use devpropdef::*;

pub use docobj::*;

pub use dpapi::*;

pub use dsgetdc::*;

pub use dsound::*;

pub use dsrole::*;

pub use dwmapi::*;

pub use dwrite::*;

pub use dxgi::*;

pub use dxgi1_2::*;

pub use dxgi1_3::*;

pub use dxgi1_4::*;

pub use dxgiformat::*;

pub use dxgitype::*;

pub use errhandlingapi::*;

pub use excpt::*;

pub use fileapi::*;

pub use gl::*;

pub use guiddef::*;

pub use heapapi::*;

pub use hidclass::*;

pub use hidpi::*;

pub use hidsdi::*;

pub use hidusage::*;

pub use hstring::*;

pub use http::*;

pub use imm::*;

pub use inaddr::*;

pub use inspectable::*;

pub use ksmedia::*;

pub use libloaderapi::*;

pub use lmaccess::*;

pub use lmcons::*;

pub use lmdfs::*;

pub use lmerrlog::*;

pub use lmjoin::*;

pub use lsalookup::*;

pub use memoryapi::*;

pub use minwinbase::*;

pub use minwindef::*;

pub use mmdeviceapi::*;

pub use mmreg::*;

pub use mmsystem::*;

pub use mscat::*;

pub use mssip::*;

pub use nb30::*;

pub use ncrypt::*;

pub use ntdef::*;

pub use ntsecapi::*;

pub use ntstatus::*;

pub use oaidl::*;

pub use objbase::*;

pub use objidl::*;

pub use objidlbase::*;

pub use olectl::*;

pub use processsnapshot::*;

pub use processthreadsapi::*;

pub use propsys::*;

pub use prsht::*;

pub use psapi::*;

pub use qos::*;

pub use reason::*;

pub use rpc::*;

pub use rpcdce::*;

pub use sapi::*;

pub use schannel::*;

pub use servprov::*;

pub use setupapi::*;

pub use shellapi::*;

pub use shellscalingapi::*;

pub use shlguid::*;

pub use shlobj::*;

pub use shobjidl::*;

pub use shtypes::*;

pub use spapidef::*;

pub use sspi::*;

pub use subauth::*;

pub use synchapi::*;

pub use sysinfoapi::*;

pub use threadpoolapi::*;

pub use timezoneapi::*;

pub use tlhelp32::*;

pub use unknwnbase::*;

pub use urlhist::*;

pub use urlmon::*;

pub use usp10::*;

pub use vadefs::*;

pub use vsbackup::*;

pub use vss::*;

pub use vsserror::*;

pub use vswriter::*;

pub use werapi::*;

pub use winbase::*;

pub use wincon::*;

pub use wincred::*;

pub use wincrypt::*;

pub use windowsx::*;

pub use windef::*;

pub use windowscodecs::*;

pub use winerror::*;

pub use winevt::*;

pub use wingdi::*;

pub use winhttp::*;

pub use winioctl::*;

pub use winnetwk::*;

pub use winnls::*;

pub use winnt::*;

pub use winscard::*;

pub use winsmcrd::*;

pub use winsock2::*;

pub use winspool::*;

pub use winsvc::*;

pub use winuser::*;

pub use ws2def::*;

pub use ws2ipdef::*;

pub use ws2spi::*;

pub use ws2tcpip::*;

pub use wtypes::*;

pub use wtypesbase::*;

pub use xinput::*;

Modules

audioclient
this ALWAYS GENERATED file contains the definitions for the interfaces

audiosessiontypes

basetsd
Type definitions for the basic sized types.

bcrypt
Cryptographic Primitive API Prototypes and Definitions

cfg
common Configuration Manager definitions for both user mode and kernel mode code

cfgmgr32
user APIs for the Configuration Manager

commctrl

commdlg
32-Bit Common Dialog APIs

corsym
Common Language Runtime Debugging Symbol Reader/Writer/Binder Interfaces

d2d1
Mappings for the contents of d2d1.h

d2dbasetypes
Mappings for the contents of d2dbasetypes.h

d3d11

d3d12

d3d12sdklayers

d3d9
Direct3D include file

d3d9caps
Direct3D capabilities include file

d3d9types
Direct3D capabilities include file

d3dcommon
Mappings for the contents of d3dcommon.h

dbghelp
DbgHelp include file

dcommon
Mappings for the contents of dcommon.h

devpropdef
Defines property types and keys for the Plug and Play Device Property API

docobj

dpapi
Data Protection API Prototypes and Definitions

dsgetdc
This file contains structures, function prototypes, and definitions for the DsGetDcName API.

dsound
DSound procedure declarations, constant definitions and macros

dsrole
Contains public interfaces to query the network roles of workstations, servers, and DCs

dwmapi
Procedure declarations, constant definitions, and macros for the NLS component.

dwrite
DirectX Typography Services public API definitions.

dxgi
Mappings for the contents of dxgi.h

dxgi1_2
Mappings for the contents of dxgi1_2.h

dxgi1_3
Mappings for the contents of dxgi1_3.h

dxgi1_4
Mappings for the contents of dxgi1_4.h

dxgiformat
Mappings for the contents of dxgiformat.h

dxgitype
Mappings for the contents of dxgitype.h

errhandlingapi
ApiSet Contract for api-ms-win-core-errorhandling-l1

excpt

fileapi
ApiSet Contract for api-ms-win-core-file-l1

gl

guiddef

heapapi
ApiSet Contract for api-ms-win-core-heap-l1

hidclass

hidpi

hidsdi

hidusage

hstring
This interface definition contains typedefs for Windows Runtime data types.

http
HTTP API specification

imm

inaddr
IPv4 Internet address

inspectable

ksmedia

libloaderapi
ApiSet Contract for api-ms-win-core-libraryloader-l1

lmaccess

lmcons
This file contains constants used throughout the LAN Manager API header files.

lmdfs

lmerrlog

lmjoin

lsalookup
LSA Policy Lookup API

memoryapi
ApiSet Contract for api-ms-win-core-memory-l1-1-0

minwinbase
This module defines the 32-Bit Windows Base APIs

minwindef
Basic Windows Type Definitions for minwin partition

mmdeviceapi
this ALWAYS GENERATED file contains the definitions for the interfaces

mmreg

mmsystem
MM procedure declarations, constant definitions and macros

mscat
Microsoft Internet Security Catalog API Prototypes and Definitions

mssip
Microsoft SIP Provider Prototypes and Definitions

nb30

ncrypt
Cryptographic API Prototypes and Definitions

ntdef
Type definitions for the basic types.

ntsecapi
This module defines the Local Security Authority APIs.

ntstatus
Constant definitions for the NTSTATUS values.

oaidl
Mappings for the contents of OAIdl.h

objbase
Component object model defintions

objidl
this ALWAYS GENERATED file contains the definitions for the interfaces

objidlbase
this ALWAYS GENERATED file contains the definitions for the interfaces

olectl
OLE Control interfaces

processsnapshot
Defines the process snapshot API

processthreadsapi

propsys

prsht
Interface for the Windows Property Sheet Pages

psapi
API Prototypes and Definitions for PSAPI.DLL

qos
QoS definitions for NDIS components.

reason

rpc

rpcdce

sapi
Mappings for the contents of sapi.h

schannel
Public Definitions for SCHANNEL Security Provider

servprov
Mappings for the contents of servprov.h

setupapi
Windows NT Setup and Device Installer services

shellapi

shellscalingapi

shlobj

shobjidl
this ALWAYS GENERATED file contains the definitions for the interfaces

shtypes
this ALWAYS GENERATED file contains the definitions for the interfaces

spapidef
Windows NT Setup and Device Installer services

sspi
Security Support Provider Interface Prototypes and structure definitions

subauth
Types and macros for Subauthentication Packages.

synchapi
ApiSet Contract for api-ms-win-core-synch-l1

sysinfoapi
ApiSet Contract for api-ms-win-core-sysinfo-l1.

threadpoolapi
ApiSet Contract for api-ms-win-core-threadpool-l1.

timezoneapi
ApiSet Contract for api-ms-win-core-timezone-l1

tlhelp32

unknwnbase
this ALWAYS GENERATED file contains the definitions for the interfaces

urlhist
Url History Interfaces

urlmon
Mappings for the contents of Urlmon.h

usp10
Uniscribe structure declarations and constant definitions

vadefs
Definitions of macro helpers used by . This is the topmost header in the CRT header lattice, and is always the first CRT header to be included, explicitly or implicitly. Therefore, this header also has several definitions that are used throughout the CRT.

vsbackup
VSS backup interfaces

vss
VSS header file

vsserror
VSS Error header file

vswriter
VSS Writer header file

werapi
Function prototypes for Windows Error Reporting (WER)

winbase
This module defines the 32-Bit Windows Base APIs

wincon
This module contains the public data structures, data types, and procedures exported by the NT console subsystem.

wincred
Authentication API Prototypes and Definitions

wincrypt
Cryptographic API Prototypes and Definitions

windef
Basic Windows Type Definitions

windowscodecs
Mappings for the contents of wincodec.h

windowsx
Macro APIs, window message crackers, and control APIs

winerror
error code definitions for the Win32 API functions

winevt
Windows Events API

wingdi
GDI procedure declarations, constant definitions and macros

winhttp
Windows HTTP Services API constant definitions and macros

winioctl
This module defines the 32-Bit Windows Device I/O control codes.

winnetwk
Standard WINNET Header File for WIN32

winnls
Procedure declarations, constant definitions, and macros for the NLS component.

winnt
This module defines the 32-Bit Windows types and constants that are defined by NT, but exposed through the Win32 API.

winscard
Data Protection API Prototypes and Definitions

winsmcrd

winsock2
definitions to be used with the WinSock 2 DLL and WinSock 2 applications.

winspool
Winspool header file

winsvc
Header file for the Service Control Manager

winuser
USER procedure declarations, constant definitions and macros

ws2def
This file contains the core definitions for the Winsock2 specification that can be used by both user-mode and kernel mode modules.

ws2ipdef

ws2spi
Definitions to be used with the WinSock service provider

ws2tcpip
WinSock2 Extension for TCP/IP protocols

wtypes
Mappings for the contents of wstypes.h

wtypesbase

xinput
XInput procedure declarations, constant definitions and macros

Structs

PROPVARIANT

ServerInformation

VALENTA

VALENTW

Enums

c_void
Type used to construct void pointers for use with C.

Constants

CLSCTX_ALL

CLSCTX_INPROC_HANDLER

CLSCTX_INPROC_SERVER

CLSCTX_LOCAL_SERVER

CLSCTX_REMOTE_SERVER

CLSCTX_SERVER

HKEY_CLASSES_ROOT

HKEY_CURRENT_CONFIG

HKEY_CURRENT_USER

HKEY_CURRENT_USER_LOCAL_SETTINGS

HKEY_DYN_DATA

HKEY_LOCAL_MACHINE

HKEY_PERFORMANCE_DATA

HKEY_PERFORMANCE_NLSTEXT

HKEY_PERFORMANCE_TEXT

HKEY_USERS

REG_MUI_STRING_TRUNCATE

RRF_NOEXPAND

RRF_RT_ANY

RRF_RT_DWORD

RRF_RT_QWORD

RRF_RT_REG_BINARY

RRF_RT_REG_DWORD

RRF_RT_REG_EXPAND_SZ

RRF_RT_REG_MULTI_SZ

RRF_RT_REG_NONE

RRF_RT_REG_QWORD

RRF_RT_REG_SZ

RRF_ZEROONFAILURE

SND_ALIAS

SND_ALIAS_ID

SND_APPLICATION

SND_ASYNC

SND_FILENAME

SND_LOOP

SND_MEMORY

SND_NODEFAULT

SND_NOSTOP

SND_NOWAIT

SND_PURGE

SND_RESOURCE

SND_RING

SND_SENTRY

SND_SYNC

SND_SYSTEM

Type Definitions

CO_MTA_USAGE_COOKIE

PServerInformation

PVALENTA

PVALENTW

REFERENCE_TIME

REGSAM

__int16

__int32

__int64

__int8

__uint16

__uint32

__uint64

__uint8

c_char

c_double

c_float

c_int

c_long

c_longlong

c_schar

c_short

c_uchar

c_uint

c_ulong

c_ulonglong

c_ushort

size_t

wchar_t

Implementations