Module winapi::minwindef

Basic Windows Type Definitions for minwin partition

Structs

FILETIME

Enums

HINSTANCE__

HKEY__

HKL__

HLSURF__

HMETAFILE__

HRGN__

HRSRC__

HSPRITE__

HSTR__

HTASK__

HWINSTA__

Constants

FALSE

MAX_PATH

TRUE

Functions

HIBYTE

HIWORD

LOBYTE

LOWORD

MAKELONG

MAKEWORD

Type Definitions

ATOM

BOOL

BYTE

DWORD

FARPROC
Pointer to probably a function with unknown type signature.

FLOAT

GLOBALHANDLE

HFILE

HGLOBAL

HINSTANCE

HKEY

HKL

HLOCAL

HLSURF

HMETAFILE

HMODULE

HRGN

HRSRC

HSPRITE

HSTR

HTASK

HWINSTA

INT

LOCALHANDLE

LPARAM

LPBOOL

LPBYTE

LPCVOID

LPDWORD

LPFILETIME

LPHANDLE

LPINT

LPLONG

LPVOID

LPWORD

LRESULT

NEARPROC
Pointer to probably a function with unknown type signature.

PBOOL

PBYTE

PDWORD

PFILETIME

PFLOAT

PHKEY

PINT

PROC
Pointer to probably a function with unknown type signature.

PSZ

PUCHAR

PUINT

PULONG

PUSHORT

PWORD

SPHANDLE

UCHAR

UINT

ULONG

USHORT

WORD

WPARAM