Module winapi::shared

Headers shared between user mode and kernel mode

Modules

basetsd
Type definitions for the basic sized types.

cfg
common Configuration Manager definitions for both user mode and kernel mode code

devpropdef
Defines property types and keys for the Plug and Play Device Property API

guiddef
GUID definition

hidusage

in6addr

inaddr
IPv4 Internet address

ktmtypes
Common types for KTM exposed at both the Nt- and Win32-layer

minwindef
Basic Windows Type Definitions for minwin partition

mstcpip
This module contains Microsoft-specific extensions to the core Winsock definitions.

ntdef
Type definitions for the basic types.

ntstatus
Constant definitions for the NTSTATUS values.

qos
QoS definitions for NDIS components.

rpc
Master include file for RPC applications.

rpcdce
This module contains the DCE RPC runtime APIs.

rpcndr

windef
Basic Windows Type Definitions

windowsx
Macro APIs, window message crackers, and control APIs

winerror
error code definitions for the Win32 API functions

ws2def
Core definitions for the Winsock2 specification

ws2ipdef
TCP/IP specific information for use by WinSock2 compatible applications.

wtypes

wtypesbase