winapi

Module um

Headers for user mode only

Modules

consoleapi

ApiSet Contract for api-ms-win-core-console-l1

errhandlingapi

ApiSet Contract for api-ms-win-core-errorhandling-l1

evntcons

fileapi

ApiSet Contract for api-ms-win-core-file-l1

gl

Headers for user mode only

handleapi

handleapi include file

lsalookup

LSA Policy Lookup API

minwinbase

This module defines the 32-Bit Windows Base APIs

ntsecapi

This module defines the Local Security Authority APIs.

processenv

profileapi

sspi

Security Support Provider Interface Prototypes and structure definitions

subauth

Types and macros for Subauthentication Packages.

timezoneapi

ApiSet Contract for api-ms-win-core-timezone-l1

wincon

This module contains the public data structures, data types, and procedures exported by the NT console subsystem.

wincontypes

This module contains the public data structures, data types, and procedures exported by the NT console subsystem.

wincred

Authentication API Prototypes and Definitions

wingdi

GDI procedure declarations, constant definitions and macros

winioctl

This module defines the 32-Bit Windows Device I/O control codes.

winnt

This module defines the 32-Bit Windows types and constants that are defined by NT, but exposed through the Win32 API.