Module winapi::um

Headers for user mode only

Modules

cfgmgr32
user APIs for the Configuration Manager

consoleapi
ApiSet Contract for api-ms-win-core-console-l1

errhandlingapi
ApiSet Contract for api-ms-win-core-errorhandling-l1

evntcons

fileapi
ApiSet Contract for api-ms-win-core-file-l1

gl
Headers for user mode only

handleapi
handleapi include file

ioapiset

libloaderapi
ApiSet Contract for api-ms-win-core-libraryloader-l1

lsalookup
LSA Policy Lookup API

minwinbase
This module defines the 32-Bit Windows Base APIs

mswsock

namedpipeapi

ntsecapi
This module defines the Local Security Authority APIs.

processthreadsapi
ApiSet Contract for api-ms-win-core-processthreads-l1

profileapi

reason

sspi
Security Support Provider Interface Prototypes and structure definitions

subauth
Types and macros for Subauthentication Packages.

synchapi
ApiSet Contract for api-ms-win-core-synch-l1

sysinfoapi
ApiSet Contract for api-ms-win-core-sysinfo-l1.

threadpoollegacyapiset

timezoneapi
ApiSet Contract for api-ms-win-core-timezone-l1

winbase
This module defines the 32-Bit Windows Base APIs

wincon
This module contains the public data structures, data types, and procedures exported by the NT console subsystem.

wincontypes
This module contains the public data structures, data types, and procedures exported by the NT console subsystem.

wincred
Authentication API Prototypes and Definitions

wingdi
GDI procedure declarations, constant definitions and macros

winioctl
This module defines the 32-Bit Windows Device I/O control codes.

winnt
This module defines the 32-Bit Windows types and constants that are defined by NT, but exposed through the Win32 API.

winreg

winsock2
Definitions to be used with the WinSock 2 DLL and WinSock 2 applications.

ws2tcpip
WinSock2 Extension for TCP/IP protocols