Module winapi::um

Headers for user mode only

Modules

cfgmgr32
user APIs for the Configuration Manager

consoleapi
ApiSet Contract for api-ms-win-core-console-l1

errhandlingapi
ApiSet Contract for api-ms-win-core-errorhandling-l1

fileapi
ApiSet Contract for api-ms-win-core-file-l1

gl
Headers for user mode only

handleapi
handleapi include file

libloaderapi
ApiSet Contract for api-ms-win-core-libraryloader-l1

minwinbase
This module defines the 32-Bit Windows Base APIs

processenv

processthreadsapi
ApiSet Contract for api-ms-win-core-processthreads-l1

profileapi

reason

sysinfoapi
ApiSet Contract for api-ms-win-core-sysinfo-l1.

timezoneapi
ApiSet Contract for api-ms-win-core-timezone-l1

winbase
This module defines the 32-Bit Windows Base APIs

wincon
This module contains the public data structures, data types, and procedures exported by the NT console subsystem.

wincontypes
This module contains the public data structures, data types, and procedures exported by the NT console subsystem.

wingdi
GDI procedure declarations, constant definitions and macros

winnt
This module defines the 32-Bit Windows types and constants that are defined by NT, but exposed through the Win32 API.

winreg