Module winapi::um

Headers for user mode only

Modules

cfgmgr32

user APIs for the Configuration Manager

combaseapi

Base Component Object Model defintions.

consoleapi

ApiSet Contract for api-ms-win-core-console-l1

d2d1

Mappings for the contents of d2d1.h

d2d1_1

Mappings for the contents of d2d1_1.h

d2d1effectauthor

d2d1effects

Mappings for the contents of d2d1effects.h

d2dbasetypes

Mappings for the contents of d2dbasetypes.h

d3d11

d3d11_1

d3d11_2

d3d11sdklayers

d3d11shader

d3d12

d3d12sdklayers

d3d12shader

d3dcommon

Mappings for the contents of d3dcommon.h

d3dcompiler

dcommon

Mappings for the contents of dcommon.h

dcomp

Mappings for the contents of dcomp.h

dcompanimation

Mappings for the contents of dcompanimation.h

documenttarget

dwrite

DirectX Typography Services public API definitions.

dxgidebug

errhandlingapi

ApiSet Contract for api-ms-win-core-errorhandling-l1

fileapi

ApiSet Contract for api-ms-win-core-file-l1

gl

Headers for user mode only

handleapi

handleapi include file

libloaderapi

ApiSet Contract for api-ms-win-core-libraryloader-l1

minwinbase

This module defines the 32-Bit Windows Base APIs

oaidl

Mappings for the contents of OAIdl.h

objidl

this ALWAYS GENERATED file contains the definitions for the interfaces

objidlbase

ocidl

oleauto

Mappings for the contents of OleAuto.h

processenv

processthreadsapi

ApiSet Contract for api-ms-win-core-processthreads-l1

profileapi

propidl

reason

synchapi

ApiSet Contract for api-ms-win-core-synch-l1

sysinfoapi

ApiSet Contract for api-ms-win-core-sysinfo-l1.

unknwnbase

winbase

This module defines the 32-Bit Windows Base APIs

wincodec

Mappings for the contents of wincodec.h

wincon

This module contains the public data structures, data types, and procedures exported by the NT console subsystem.

wincontypes

This module contains the public data structures, data types, and procedures exported by the NT console subsystem.

wingdi

GDI procedure declarations, constant definitions and macros

winnt

This module defines the 32-Bit Windows types and constants that are defined by NT, but exposed through the Win32 API.

winreg

winsock2

Definitions to be used with the WinSock 2 DLL and WinSock 2 applications.

winuser

USER procedure declarations, constant definitions and macros

ws2tcpip

WinSock2 Extension for TCP/IP protocols