Module winapi::um

Headers for user mode only

Modules

accctrl

aclapi

cfgmgr32

user APIs for the Configuration Manager

cguid

commapi

commctrl

commoncontrols

consoleapi

ApiSet Contract for api-ms-win-core-console-l1

errhandlingapi

ApiSet Contract for api-ms-win-core-errorhandling-l1

fileapi

ApiSet Contract for api-ms-win-core-file-l1

gl

Headers for user mode only

handleapi

handleapi include file

libloaderapi

ApiSet Contract for api-ms-win-core-libraryloader-l1

minwinbase

This module defines the 32-Bit Windows Base APIs

namedpipeapi

processenv

processthreadsapi

ApiSet Contract for api-ms-win-core-processthreads-l1

prsht

Interface for the Windows Property Sheet Pages

reason

setupapi

Public header file for Windows NT Setup and Device Installer services Dll.

spapidef

Public header file for Windows NT Setup and Device Installer services Dlls

synchapi

ApiSet Contract for api-ms-win-core-synch-l1

threadpoollegacyapiset

unknwnbase

winbase

This module defines the 32-Bit Windows Base APIs

wincon

This module contains the public data structures, data types, and procedures exported by the NT console subsystem.

wincontypes

This module contains the public data structures, data types, and procedures exported by the NT console subsystem.

wingdi

GDI procedure declarations, constant definitions and macros

winnt

This module defines the 32-Bit Windows types and constants that are defined by NT, but exposed through the Win32 API.

winreg

winsock2

Definitions to be used with the WinSock 2 DLL and WinSock 2 applications.

winuser

USER procedure declarations, constant definitions and macros

ws2tcpip

WinSock2 Extension for TCP/IP protocols